requirements.txt
asgiref==3.3.1
certifi==2020.11.8
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1.3
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
importlib-metadata==2.0.0
Markdown==3.3.3
psycopg2==2.8.6
pytz==2020.4
requests==2.24.0
sqlparse==0.4.1
urllib3==1.25.11
whitenoise==5.2.0
zipp==3.4.0
